Combined the overly used methods for checking username and password to get rid of the duplicate codes.
